+++ returns/functions.py
@@ -30,3 +30,18 @@ def safe(function):
         except Exception as exc:
             return Failure(exc)
     return decorator
+
+
+def pipeline(function):
+    """
+    Decorator to enable 'do-notation' context.
+
+    Should be used for series of computations that rely on ``.unwrap`` method.
+    """
+    @wraps(function)
+    def decorator(*args, **kwargs):
+        try:
+            return function(*args, **kwargs)
+        except UnwrapFailedError as exc:
+            return exc.halted_monad
+    return decorator