+++ tests/test_result/test_result_map.py
@@ -3,21 +3,21 @@
 from returns.result import Failure, Success
 
 
-def test_fmap_success():
+def test_map_success():
     """Ensures that Failure identity works for Success monad."""
-    assert Success(5).fmap(str) == Success('5')
+    assert Success(5).map(str) == Success('5')
 
 
-def test_fmap_failure():
+def test_map_failure():
     """Ensures that Failure identity works for Success monad."""
-    assert Failure(5).fmap(str) == Failure(5)
+    assert Failure(5).map(str) == Failure(5)
 
 
-def test_efmap_success():
+def test_fix_success():
     """Ensures that Failure identity works for Success monad."""
-    assert Success(5).efmap(str) == Success(5)
+    assert Success(5).fix(str) == Success(5)
 
 
-def test_efmap_failure():
+def test_fix_failure():
     """Ensures that Failure identity works for Success monad."""
-    assert Failure(5).efmap(str) == Success('5')
+    assert Failure(5).fix(str) == Success('5')